idlemapp is index in g_idlebuf
Currently idlemapp is first being used as the byte index of the idlemap
byte for the PFN because it is multipled by BITMAP_CHUNK_SIZE.
This is useful for comparing to the size of g_idlebuf for bounds
checking.
Later it is used as an *index* which is an error because the index value
is the number of data chunks (in this case uint64) to jump in the
buffer.

The fix here makes idlemapp *always* the index in words, doing an
explicit multiply by the size of a word for bounds checking.