Increment pointer by words read, not bytes
When a value is added to a pointer, the address of the pointer is
incremented by the size of the data being pointed to multipled by the
value being added.
The read() syscall returns the number of bytes read, hence we need to
divide by the size of a word to get the amount to increment the pointer
by.