Log accesses bits for virtual pages to file
This is collecting all of the access bits for each virtual page in
order, and logging each contiguous range of addresses to a file on disk.
One bit per page, so it's quite small.

The purpose of this is to support running additional tooling on top of
the log results to generate page activity maps over time.