Add mem_analyse util
This tool looks at the memory dumped by memdump and generates stats
about the data in the memory pages.
The stats include number of pages, number of zero pages, pages with
repeating patterns and how many times a particular page content occurs.
The objective is to provide visibility into the content of pages that
can be used for compression or KSM techniques.