Read process memory in contiguos blocks
The previous approach of reading one page at a time was simple but
inefficient. The complexity in trying to do a big read is that reading
a non-resident page will cause it to become resident. This solution
looks at blocks of pages which are already resident and reads the block
as a whole.

A proper logger is also introduced, and dump supports variable sleep
time. (this was a messy commit, sorry).

WIP read mem ranges

Clean up logging

Fix memory dump start address

Initially each bulk read started at the beginning of the segment. Oops.
Fix it to start at the page offset.

Also some log percentages and handle checking for swapping.