Return 64-bit flags for process and PFN
We'd like to capture the whole 64-bit page flag for in-guest tracking,
so to keep the format the same we'll also capture the 64-bit proc
pagemap word.
Doing some bit mapping to put the idle bit and zero bit in the same
place for both formats to simplify post processing.