Support inspecting host memory
Those host needs a compile time flag to allow reading of /dev/mem, so
it'll be kind of rare that we can do it. But the functionality exists if
we want it.
One bug and one problem: bug is that we're looking at contiguous PFNs;
this is wrong; we must get PFNs from /proc/iomem. Problem is that each
page is being read individually from /dev/mem. Need to batch and maybe
only look at in-use pages like we do for process mem.